+++ src/interfaces.ts
@@ -1,4 +1,4 @@
-import { Request } from 'express';
+import { Request, Response } from 'express';
 import * as pino from 'pino';
 
 interface LoggerOptions extends pino.LoggerOptions {
@@ -6,20 +6,23 @@ interface LoggerOptions extends pino.LoggerOptions {
 }
 
 // this is basically enhanced version of pino-multi-stream.Streams type
-export interface AckeeLoggerStream {
+export interface CosmasStream {
     level?: pino.LevelWithSilent;
     maxLevel?: number; // this is not processed by pino, so we need the number directly
     stream: NodeJS.WritableStream;
 }
 
-export interface AckeeLoggerOptions {
+export interface CosmasOptions {
     disableFields?: string[];
     enableFields?: string[];
     defaultLevel?: pino.LevelWithSilent;
     disableStackdriverFormat?: boolean;
-    streams?: AckeeLoggerStream[];
+    streams?: CosmasStream[];
     ignoredHttpMethods?: string[];
     config?: LoggerOptions;
     pretty?: boolean;
-    skip?: (req: Request) => boolean;
+    sentry?: string | boolean;
+    sentryLevel?: pino.LevelWithSilent;
+    loggerName?: string;
+    skip?: (req: Request, res?: Response) => boolean;
 }