+++ src/streams.ts
@@ -2,41 +2,46 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as pino from 'pino';
 import { Transform, TransformCallback } from 'stream';
-import { AckeeLoggerOptions, AckeeLoggerStream } from './interfaces';
+import * as util from 'util';
+import { loggerNameKey, pkgVersionKey } from './index';
+import { CosmasOptions, CosmasStream } from './interfaces';
 import { levels } from './levels';
 import { StackDriverFormatStream } from './stackdriver';
 
-const isString = (x: any) => typeof x === 'string' || x instanceof String;
-
 const pkgJson = JSON.parse(fs.readFileSync(path.resolve(path.join(__dirname, '..', 'package.json')), 'utf8'));
 
-const getDefaultTransformStream = (options: AckeeLoggerOptions & { messageKey: string; loggerName?: string }) => {
+const getDefaultTransformStream = (options: CosmasOptions & { messageKey: string; loggerName?: string }) => {
     class DefaultTransformStream extends Transform {
         // tslint:disable-next-line:function-name
         public _transform(chunk: any, _encoding: string, callback: TransformCallback) {
             const obj = JSON.parse(chunk);
             const loggerName = options.loggerName;
-            if (options.pretty) {
-                obj['name\0'] = obj.name; // add null character so that it is not interpreted by pino-pretty but still visible to user unchanged
-                delete obj.name;
-                if (loggerName) {
-                    obj.name = loggerName;
-                }
-            } else {
-                obj.pkgVersion = pkgJson.version;
-                if (obj[options.messageKey] && isString(obj[options.messageKey]) && loggerName) {
-                    obj[options.messageKey] = `[${loggerName}] ${obj[options.messageKey]}`;
-                }
+            let res;
+            if (loggerName) {
+                // always put logger name to message
+                obj[options.messageKey] = `[${loggerName}] ${obj[options.messageKey]}`;
+            }
+            if (loggerName && !options.pretty) {
+                // do not put logger name field to pretty outputs
+                obj[loggerNameKey] = loggerName;
             }
 
-            this.push(`${JSON.stringify(obj)}\n`);
+            if (options.pretty) {
+                res = util.inspect(obj, { colors: true, showHidden: true, depth: Infinity });
+            } else {
+                // do not put pkgVersion to pretty outputs
+                obj[pkgVersionKey] = pkgJson.version;
+                res = JSON.stringify(obj);
+            }
+
+            this.push(`${res}\n`);
             callback();
         }
     }
     return DefaultTransformStream;
 };
 
-const decorateStreams = <T extends Transform>(streams: AckeeLoggerStream[], streamClass: new () => T) => {
+const decorateStreams = <T extends Transform>(streams: CosmasStream[], streamClass: new () => T) => {
     return streams.map(stream => {
         const newStream = new streamClass();
         newStream.pipe(stream.stream);
@@ -50,39 +55,27 @@ const decorateStreams = <T extends Transform>(streams: AckeeLoggerStream[], stre
 
 const initLoggerStreams = (
     defaultLevel: pino.LevelWithSilent,
-    options: AckeeLoggerOptions & { messageKey: string; loggerName?: string }
+    options: CosmasOptions & { messageKey: string; loggerName?: string }
 ) => {
-    let streams: AckeeLoggerStream[];
+    let streams: CosmasStream[];
     if (options.streams) {
         streams = options.streams.map(stream => Object.assign({ level: defaultLevel }, stream));
-    } else if (options.pretty) {
-        const pretty = pino({
-            prettyPrint: {
-                levelFirst: true,
-            },
-        });
-        pretty.pipe(process.stdout);
-        const prettyErr = pino({
-            prettyPrint: {
-                levelFirst: true,
-            },
-        });
-        prettyErr.pipe(process.stderr);
-        streams = [
-            { level: defaultLevel, maxLevel: levels.warn, stream: pretty as any },
-            { level: 'warn', stream: prettyErr as any },
-        ];
     } else {
         streams = [
             { level: defaultLevel, maxLevel: levels.warn, stream: process.stdout },
             { level: 'warn', stream: process.stderr },
         ];
     }
-    if (!options.disableStackdriverFormat) {
+    if (!options.pretty && !options.disableStackdriverFormat) {
         streams = decorateStreams(streams, StackDriverFormatStream);
     }
 
     streams = decorateStreams(streams, getDefaultTransformStream(options));
+
+    if (options.sentry) {
+        const { createSentryTransformStream } = require('./sentry');
+        streams = decorateStreams(streams, createSentryTransformStream(options));
+    }
 
     return streams;
 };